Upgrade to VS2022, reformatted the code, addressed two SAST findings

+++ ReparsePoint.hpp
@@ -13,7 +13,7 @@
 #ifndef __REPARSEPOINT_H_VER__
 #define __REPARSEPOINT_H_VER__ 2017091820
 #if defined(_MSC_VER) && (_MSC_VER >= 1020)
-#pragma once
+#    pragma once
 #endif // Check for "#pragma once" support
 
 #include <tchar.h>
@@ -23,66 +23,66 @@
 #include "ReparseTags.h"
 #include "FileAttributes.h"
 
-#define WIN32_UNICODE_PREFIX                    L"\\\\?\\"
+#define WIN32_UNICODE_PREFIX L"\\\\?\\"
 
 #ifndef SYMLINK_FLAG_FULLPATH
 /* The substitute name is a full path name. */
-#define SYMLINK_FLAG_FULLPATH 0x00000000
+#    define SYMLINK_FLAG_FULLPATH 0x00000000
 #endif // SYMLINK_FLAG_FULLPATH
 
 #ifndef SYMLINK_FLAG_RELATIVE
 /* The substitute name is a path name relative to the directory containing the symbolic
 link. */
-#define SYMLINK_FLAG_RELATIVE 0x00000001
+#    define SYMLINK_FLAG_RELATIVE 0x00000001
 #endif // SYMLINK_FLAG_RELATIVE
 
 class CReparsePoint
 {
     // Allow the header to be included again outside of this file (and class)
     // Still this header needs to be included before anyone else who requires VerySimpleBuf.hpp
-#   define __VERYSIMPLEBUF_MULTI_INC__
-#   include "VerySimpleBuf.hpp"
-#   undef __VERYSIMPLEBUF_HPP_VER__
+#define __VERYSIMPLEBUF_MULTI_INC__
+#include "VerySimpleBuf.hpp"
+#undef __VERYSIMPLEBUF_HPP_VER__
     // Just make it more readable
     typedef CVerySimpleBuf<WCHAR> CWideString, *PWideString;
-    typedef CVerySimpleBuf<CHAR>  CAnsiString, *PAnsiString;
+    typedef CVerySimpleBuf<CHAR> CAnsiString, *PAnsiString;
     typedef CVerySimpleBuf<TCHAR> CTString, *PTString;
-    typedef CVerySimpleBuf<BYTE>  ByteBuf;
+    typedef CVerySimpleBuf<BYTE> ByteBuf;
 
-#   include <PshPack1.h>
+#include <PshPack1.h>
     // The Microsoft version of the reparse data, as opposed to REPARSE_GUID_DATA_BUFFER
     typedef struct _REPARSE_DATA_BUFFER
     {
-        ULONG  ReparseTag;
-        USHORT  ReparseDataLength;
-        USHORT  Reserved;
+        ULONG ReparseTag;
+        USHORT ReparseDataLength;
+        USHORT Reserved;
         union
         {
             struct
             {
-                USHORT  SubstituteNameOffset;
-                USHORT  SubstituteNameLength;
-                USHORT  PrintNameOffset;
-                USHORT  PrintNameLength;
-                ULONG   Flags;
-                WCHAR   PathBuffer[1];
+                USHORT SubstituteNameOffset;
+                USHORT SubstituteNameLength;
+                USHORT PrintNameOffset;
+                USHORT PrintNameLength;
+                ULONG Flags;
+                WCHAR PathBuffer[1];
             } SymbolicLinkReparseBuffer;
             struct
             {
-                USHORT  SubstituteNameOffset;
-                USHORT  SubstituteNameLength;
-                USHORT  PrintNameOffset;
-                USHORT  PrintNameLength;
-                WCHAR   PathBuffer[1];
+                USHORT SubstituteNameOffset;
+                USHORT SubstituteNameLength;
+                USHORT PrintNameOffset;
+                USHORT PrintNameLength;
+                WCHAR PathBuffer[1];
             } MountPointReparseBuffer;
             struct
             {
-                UCHAR  DataBuffer[1];
+                UCHAR DataBuffer[1];
             } GenericReparseBuffer;
         };
     } REPARSE_DATA_BUFFER, *PREPARSE_DATA_BUFFER;
-#   include <PopPack.h>
-public:
+#include <PopPack.h>
+  public:
     CReparsePoint(WCHAR const* Path)
         : m_LastError(ERROR_SUCCESS)
         , m_Path(normalizePath_(Path))
@@ -207,7 +207,7 @@ public:
     }
 #endif // RP_QUERY_FILE_ID
 
-private:
+  private:
     CVerySimpleBuf<WCHAR> normalizePath_(WCHAR const* Path)
     {
         if (Path)
@@ -222,7 +222,7 @@ private:
                     DWORD dwNeeded = ::GetCurrentDirectoryW(0, NULL);
                     if (dwNeeded)
                     {
-                        if (sPath.reAlloc(1 + dwNeeded + sPath.getCount()))
+                        if (sPath.reAlloc((size_t)1 + dwNeeded + sPath.getCount()))
                         {
                             if (0 < ::GetCurrentDirectoryW(static_cast<DWORD>(sPath.getCount() - sPath.getCountZ()), sPath.getBuf() + sPath.getCountZ()))
                             {
@@ -235,7 +235,7 @@ private:
                 LPWSTR filePart = 0;
                 // dummy call to evaluate required length
                 DWORD dwNeeded = ::GetFullPathNameW(Path, 0, sPath.getBuf(), &filePart);
-                if (sPath.reAlloc(1 + dwNeeded + wcslen(Path)))
+                if (sPath.reAlloc((size_t)1 + dwNeeded + wcslen(Path)))
                 {
                     if (0 < ::GetFullPathNameW(Path, static_cast<DWORD>(sPath.getCount() - sPath.getCountZ()), sPath.getBuf() + sPath.getCountZ(), &filePart))
                     {
@@ -258,7 +258,7 @@ private:
             if (isReparsePoint())
             {
 #ifdef RP_QUERY_FILE_ID
-                BY_HANDLE_FILE_INFORMATION bhfi = { 0 };
+                BY_HANDLE_FILE_INFORMATION bhfi = {0};
                 if (::GetFileInformationByHandle(hFile, &bhfi))
                 {
                     m_FileId.HighPart = bhfi.nFileIndexHigh;
@@ -274,15 +274,7 @@ private:
                 DWORD dwReturned = 0;
                 if (buf.reAlloc(MAXIMUM_REPARSE_DATA_BUFFER_SIZE + REPARSE_GUID_DATA_BUFFER_HEADER_SIZE))
                 {
-                    if (::DeviceIoControl(
-                        hFile,
-                        FSCTL_GET_REPARSE_POINT,
-                        NULL,
-                        0,
-                        buf.getBuf(),
-                        static_cast<DWORD>(buf.getByteCount()),
-                        &dwReturned,
-                        0))
+                    if (::DeviceIoControl(hFile, FSCTL_GET_REPARSE_POINT, NULL, 0, buf.getBuf(), static_cast<DWORD>(buf.getByteCount()), &dwReturned, 0))
                     {
                         PREPARSE_GUID_DATA_BUFFER const repbuf = reinterpret_cast<PREPARSE_GUID_DATA_BUFFER>(buf.getBuf());
                         m_ReparseTag = repbuf->ReparseTag;
@@ -290,25 +282,33 @@ private:
                         if (0 != IsReparseTagMicrosoft(m_ReparseTag))
                         {
                             PREPARSE_DATA_BUFFER const msrepbuf = reinterpret_cast<PREPARSE_DATA_BUFFER>(repbuf);
-                            if ((IO_REPARSE_TAG_SYMLINK == m_ReparseTag)
-                                && (m_PrintName.reAlloc(msrepbuf->SymbolicLinkReparseBuffer.PrintNameLength / sizeof(WCHAR) + 1))
-                                && (m_SubstName.reAlloc(msrepbuf->SymbolicLinkReparseBuffer.SubstituteNameLength / sizeof(WCHAR) + 1))
-                                )
+                            if ((IO_REPARSE_TAG_SYMLINK == m_ReparseTag) &&
+                                (m_PrintName.reAlloc(msrepbuf->SymbolicLinkReparseBuffer.PrintNameLength / sizeof(WCHAR) + 1)) &&
+                                (m_SubstName.reAlloc(msrepbuf->SymbolicLinkReparseBuffer.SubstituteNameLength / sizeof(WCHAR) + 1)))
                             {
-                                memcpy(m_PrintName.getBuf(), &msrepbuf->SymbolicLinkReparseBuffer.PathBuffer[msrepbuf->SymbolicLinkReparseBuffer.PrintNameOffset / sizeof(WCHAR)], msrepbuf->SymbolicLinkReparseBuffer.PrintNameLength);
-                                memcpy(m_SubstName.getBuf(), &msrepbuf->SymbolicLinkReparseBuffer.PathBuffer[msrepbuf->SymbolicLinkReparseBuffer.SubstituteNameOffset / sizeof(WCHAR)], msrepbuf->SymbolicLinkReparseBuffer.SubstituteNameLength);
+                                memcpy(m_PrintName.getBuf(),
+                                       &msrepbuf->SymbolicLinkReparseBuffer.PathBuffer[msrepbuf->SymbolicLinkReparseBuffer.PrintNameOffset / sizeof(WCHAR)],
+                                       msrepbuf->SymbolicLinkReparseBuffer.PrintNameLength);
+                                memcpy(
+                                    m_SubstName.getBuf(),
+                                    &msrepbuf->SymbolicLinkReparseBuffer.PathBuffer[msrepbuf->SymbolicLinkReparseBuffer.SubstituteNameOffset / sizeof(WCHAR)],
+                                    msrepbuf->SymbolicLinkReparseBuffer.SubstituteNameLength);
                             }
-                            else if ((IO_REPARSE_TAG_MOUNT_POINT == m_ReparseTag)
-                                && (m_PrintName.reAlloc(msrepbuf->MountPointReparseBuffer.PrintNameLength / sizeof(WCHAR) + 1))
-                                && (m_SubstName.reAlloc(msrepbuf->MountPointReparseBuffer.SubstituteNameLength / sizeof(WCHAR) + 1))
-                                )
+                            else if ((IO_REPARSE_TAG_MOUNT_POINT == m_ReparseTag) &&
+                                     (m_PrintName.reAlloc(msrepbuf->MountPointReparseBuffer.PrintNameLength / sizeof(WCHAR) + 1)) &&
+                                     (m_SubstName.reAlloc(msrepbuf->MountPointReparseBuffer.SubstituteNameLength / sizeof(WCHAR) + 1)))
                             {
-                                memcpy(m_PrintName.getBuf(), &msrepbuf->MountPointReparseBuffer.PathBuffer[msrepbuf->MountPointReparseBuffer.PrintNameOffset / sizeof(WCHAR)], msrepbuf->MountPointReparseBuffer.PrintNameLength);
-                                memcpy(m_SubstName.getBuf(), &msrepbuf->MountPointReparseBuffer.PathBuffer[msrepbuf->MountPointReparseBuffer.SubstituteNameOffset / sizeof(WCHAR)], msrepbuf->MountPointReparseBuffer.SubstituteNameLength);
+                                memcpy(m_PrintName.getBuf(),
+                                       &msrepbuf->MountPointReparseBuffer.PathBuffer[msrepbuf->MountPointReparseBuffer.PrintNameOffset / sizeof(WCHAR)],
+                                       msrepbuf->MountPointReparseBuffer.PrintNameLength);
+                                memcpy(m_SubstName.getBuf(),
+                                       &msrepbuf->MountPointReparseBuffer.PathBuffer[msrepbuf->MountPointReparseBuffer.SubstituteNameOffset / sizeof(WCHAR)],
+                                       msrepbuf->MountPointReparseBuffer.SubstituteNameLength);
                             }
                             m_SubstNameCanonical = m_SubstName;
                             // If the prefix is of the native NT form, change it to the Win32 form instead
-                            if (m_SubstNameCanonical.getCountZ() && (0 == wcsncmp(NativePrefix, m_SubstNameCanonical.getBuf(), sizeof(NativePrefix) / sizeof(NativePrefix[0]) - 1)))
+                            if (m_SubstNameCanonical.getCountZ() &&
+                                (0 == wcsncmp(NativePrefix, m_SubstNameCanonical.getBuf(), sizeof(NativePrefix) / sizeof(NativePrefix[0]) - 1)))
                             {
                                 m_SubstNameCanonical.getBuf()[1] = L'\\';
                             }
@@ -319,7 +319,7 @@ private:
 #ifdef RP_QUERY_FILE_ID
             else
             {
-                BY_HANDLE_FILE_INFORMATION bhfi = { 0 };
+                BY_HANDLE_FILE_INFORMATION bhfi = {0};
                 if (::GetFileInformationByHandle(hFile, &bhfi))
                 {
                     m_FileId.HighPart = bhfi.nFileIndexHigh;
@@ -339,18 +339,18 @@ private:
         }
     }
 
-    LONG                    m_LastError;
-    CVerySimpleBuf<WCHAR>   m_Path;
-    CVerySimpleBuf<WCHAR>   m_SubstName;
-    CVerySimpleBuf<WCHAR>   m_PrintName;
-    CVerySimpleBuf<WCHAR>   m_SubstNameCanonical;
-    GUID                    m_ReparseGuid;
-    DWORD const             m_Attr;
-    DWORD const             m_OpenFlags;
-    DWORD                   m_ReparseTag;
+    LONG m_LastError;
+    CVerySimpleBuf<WCHAR> m_Path;
+    CVerySimpleBuf<WCHAR> m_SubstName;
+    CVerySimpleBuf<WCHAR> m_PrintName;
+    CVerySimpleBuf<WCHAR> m_SubstNameCanonical;
+    GUID m_ReparseGuid;
+    DWORD const m_Attr;
+    DWORD const m_OpenFlags;
+    DWORD m_ReparseTag;
     CVerySimpleBuf<unsigned char> m_RawReparseData;
 #ifdef RP_QUERY_FILE_ID
-    ULARGE_INTEGER          m_FileId;
+    ULARGE_INTEGER m_FileId;
 #endif // RP_QUERY_FILE_ID
     // Hide these
     CReparsePoint();
